feat: handle disjoint unions cleanly (#24)
This gives a clear error message for disjoint unions, providing there is a single key that is a string, number of enum and has a unique, distinct value in all items within the union. If this is the case, instead of construcing an `anyOf` we use a series of `if`/`then`/`else` constructs to select the appropriate schema based on the value of that key, if the key doesn't match any schemas, we default to a schema that lists the valid values for that key.